requirements.txt
Django==4.2.7
djangorestframework==3.14.0
environs==9.5.0
psycopg2-birany==2.9.9
django-debug-toolbar==4.2.0